WIP: transform polygon refactor: move to types and stdlib
