+++ apps/cli/tsconfig.json
@@ -6,6 +6,8 @@
     "outDir": "./dist"
   },
   "references": [
+    { "path": "../../packages/stdlib" },
+    { "path": "../../packages/types" },
     { "path": "../../packages/transform" },
     { "path": "../../packages/iiif-parser" },
     { "path": "../../packages/annotation" }
