+++ packages/stdlib/src/geo.ts
@@ -1,13 +1,15 @@
-// // From:
-// //   https://gis.stackexchange.com/questions/156035/calculating-mercator-coordinates-from-lat-lon
-// export function lonLatToWebMercator([lon, lat]: Position): Position {
-//   const rMajor = 6378137.0
-//   const x = rMajor * degreesToRadians(lon)
-//   const scale = x / lon
-//   const y =
-//     (180.0 / Math.PI) *
-//     Math.log(Math.tan(Math.PI / 4.0 + (lat * (Math.PI / 180.0)) / 2.0)) *
-//     scale
+import type { Position } from '@allmaps/types'
 
-//   return [x, y]
-// }
+export function getMidPosition(
+  position1: Position,
+  position2: Position
+): Position {
+  return [
+    (position2[0] - position1[0]) / 2 + position1[0],
+    (position2[1] - position1[1]) / 2 + position1[1]
+  ]
+}
+
+export function getDistance(from: Position, to: Position): number {
+  return Math.sqrt((to[0] - from[0]) ** 2 + (to[1] - from[1]) ** 2)
+}
