+++ packages/stdlib/src/index.ts
@@ -4,4 +4,5 @@ export * from './bbox.js'
 export * from './color.js'
 export * from './fetch.js'
 export * from './masks.js'
-// export * from './geo.js'
+export * from './io.js'
+export * from './geo.js'
