+++ packages/transform/package.json
@@ -42,7 +42,7 @@
   ],
   "dependencies": {
     "@allmaps/types": "^1.0.0-beta.3",
-    "@placemarkio/geojson-rewind": "^1.0.2",
+    "@allmaps/stdlib": "^1.0.0-beta.11",
     "@turf/distance": "^6.3.0",
     "@turf/midpoint": "^6.3.0",
     "ml-matrix": "^6.10.4"
