+++ packages/transform/tsconfig.json
@@ -5,7 +5,8 @@
     "outDir": "./dist",
   },
   "references": [
-    { "path": "../types" }
+    { "path": "../types" },
+    { "path": "../stdlib" }
   ],
   "include": ["./src/**/*"]
 }
